added vars for constructor relating to graphic option additions in the main menu
Changed start method implementation to use gameStage as parameter though did already from the constructor calling start(gameStage)
but the refractor makes this more clear.

removed the stage background color wasted memory/useless code

Added Circle[] circle = new Circle[(bandRate / 5)] for circle visualization this must be done prior to the if state that checks that
the circles should or should not be displayed based on the user graphic selection.
also the code for the circles was replaced.

Finally the whole document was got formated for tabs and spacing via Netbeans format operation for readability purposes.